+++ frictionless/plugins/multipart.py
@@ -5,7 +5,6 @@ from ..plugin import Plugin
 from ..loader import Loader
 from ..system import system
 from .. import helpers
-from .. import config
 
 
 # NOTE:
@@ -66,7 +65,7 @@ class MultipartControl(Control):
 
     @property
     def chunk_size(self):
-        return self.get("chunkSize", config.DEFAULT_MULTIPART_CHUNK_SIZE)
+        return self.get("chunkSize", DEFAULT_CHUNK_SIZE)
 
     # Expand
 
@@ -123,6 +122,9 @@ class MultipartLoader(Loader):
 
 
 # Internal
+
+
+DEFAULT_CHUNK_SIZE = 100000000
 
 
 class MultipartByteStream: